Homework 17. Репозитории и сервисы готовы и ресурсы написаны. После перенесения ресурсов в api модуль проект перестал запускаться. Добавил первоначальный Changelog из 14 ДЗ. Сделал второй с изменениями.

+++ homework-impl/src/main/java/com/epam/igushkin/homework/repository/ProductRepository.java
@@ -4,6 +4,9 @@ import com.epam.igushkin.homework.domain.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+/**
+ * Интерфейс для работы с репозиторием и выполения CRUD операций.
+ */
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 }
