+++ homework-impl/src/main/java/com/epam/igushkin/homework/services/ProductService.java
@@ -1,21 +1,21 @@
 package com.epam.igushkin.homework.services;
 
-import com.epam.igushkin.homework.domain.entity.Order;
 import com.epam.igushkin.homework.domain.entity.Product;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+/**
+ * Описывает методы для сохранения, получения, удаления, изменения продукта.
+ */
 public interface ProductService {
     Product save(Product product);
 
     List<Product> getAll();
 
-    Product findById(Long id);
+    Product findById(Integer id);
 
-    Product update (Long id, Product product);
+    Product update (Integer id, Product product);
 
-    boolean delete (Long id);
+    boolean delete (Integer id);
 
 }
