+++ homework-impl/src/main/java/com/epam/igushkin/homework/services/impl/ProductServiceImpl.java
@@ -17,24 +17,48 @@ public class ProductServiceImpl implements ProductService {
 
     private final ProductRepository productRepository;
 
+    /**
+     * Сохраняет продукт в репозитроий.
+     *
+     * @param product продукт для сохранения.
+     * @return сохраненный продукт.
+     */
     @Override
     public Product save(Product product) {
         return productRepository.save(product);
     }
 
+    /**
+     * Возвращает все продукты из репозитория.
+     *
+     * @return список продуктов.
+     */
     @Override
     public List<Product> getAll() {
         return productRepository.findAll();
     }
 
+    /**
+     * Находит продукт по его id в репозитории.
+     *
+     * @param id уникальный номер продукта.
+     * @return найденный продукт.
+     */
     @Override
-    public Product findById(Long id) {
+    public Product findById(Integer id) {
         var productOptional = productRepository.findById(id);
         return productOptional.orElseThrow(() -> new NoEntityFoundException("Невозможно найти Продукт с номером " + id));
     }
 
+    /**
+     * Обновляет данные продукт в репозитории.
+     *
+     * @param id уникальный номер продукта.
+     * @param product данные для обновления продукта.
+     * @return обновленный продукт.
+     */
     @Override
-    public Product update(Long id, Product product) {
+    public Product update(Integer id, Product product) {
         var oldProductOpt = productRepository.findById(id);
         if (oldProductOpt.isEmpty()) {
             throw new NoEntityFoundException("Заказ с id" + id + "не найден.");
@@ -48,8 +72,14 @@ public class ProductServiceImpl implements ProductService {
         return productRepository.save(product);
     }
 
+    /**
+     * Удаляет продукт из репозитория по id.
+     *
+     * @param id уникальный номер продукта.
+     * @return результат удаления (true или false).
+     */
     @Override
-    public boolean delete(Long id) {
+    public boolean delete(Integer id) {
         boolean success = false;
         if (productRepository.existsById(id)) {
             productRepository.delete(productRepository.findById(id).get());
